rasabot/requirements.txt
africastalking==1.2.7
Django==5.0.1
Flask==3.0.3
python-dotenv==1.0.1
rasa==3.6.20
tensorflow
ujson==5.10.0
openai==1.35.7
